Add multiple liquid interface support
Now allows for expansion of the amount of fluids that we use as inputs.

This required treating the endertank as a modem connected peripheral to make it a valid transfer location. (looks like native cc is kinda like transposers in that it only lets you send to sides local to the perpherial)